Add Requestor Request and Error to zod schemas
