+++ fpx/src/lib.rs
@@ -1,1 +1,5 @@
+pub mod api;
+pub mod data;
+pub mod events;
+pub mod inspector;
 pub mod models;
